Decimate input samples to max 24kHz ...
Benchmarking time measurement reduces from 4468ms to 2160ms.

+++ libraries/lib-music-information-retrieval/MirAudioReader.h
@@ -17,7 +17,7 @@ namespace MIR
 class MirAudioReader
 {
 public:
-   virtual int GetSampleRate() const = 0;
+   virtual double GetSampleRate() const = 0;
    virtual long long GetNumSamples() const = 0;
    virtual void
    ReadFloats(float* buffer, long long where, size_t numFrames) const = 0;
